feat: 네비게이션 버튼과 아코디언에 active 상태 반영
추가 작업:
- Sheet 컴포넌트 파일 분리
- 사이트 제목은 metadata 파일에서 관리
