+++ apps/yjshd/lib/metadata.ts
@@ -1,0 +1,1 @@
+export const PAGE_H1 = "연세정성내과의원";
