+++ packages/common/src/lib/uri.ts
@@ -1,0 +1,3 @@
+export const decodeURIS = (...uris: (string | undefined)[]) => {
+  return uris.map((uri) => uri && decodeURI(uri));
+};
