+++ packages/common/src/structure/params.types.ts
@@ -1,3 +1,4 @@
+export type DefaultParams = { category?: string; subCategory?: string; slug?: string };
 export type CategoryParams = { category: string; subCategory?: string; slug?: string };
 export type SubCategoryParams = { category: string; subCategory: string; slug?: string };
 export type SlugParams = { category: string; subCategory: string; slug: string };
